# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vamo al Game API",
    "version": "1.0.0",
    "description": "Contrato base para la arquitectura dual web + movil."
  },
  "servers": [
    {
      "url": "http://localhost/api/v1",
      "description": "Entorno local por defecto"
    }
  ],
  "paths": {
...
    "/command-center/settings": {
      "get": {
        "operationId": "commandCenterSettings",
        "summary": "Branding del centro de mando",
        "security": [
          {
            "sanctumBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Configuracion cargada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandCenterSettingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sin permisos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "commandCenterUpdateSettings",
        "summary": "Actualiza logo y favicon del centro de mando",
        "security": [
          {
            "sanctumBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/BrandingUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branding actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandCenterSettingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sin permisos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Datos invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/settings/profile": {
      "patch": {
        "operationId": "updateProfile",
        "summary": "Actualiza nombre y correo del usuario autenticado",
        "security": [
          {
            "sanctumBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Datos invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteAccount",
        "summary": "Elimina la cuenta autenticada",
        "security": [
          {
            "sanctumBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cuenta eliminada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptySuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Password invalida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sanctumBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token"
      }
    },
    "schemas": {
...
      "BrandingUpdateRequest": {
        "type": "object",
        "properties": {
          "logo": {
            "type": [
              "string",
              "null"
            ],
            "format": "binary"
          },
          "favicon": {
            "type": [
              "string",
              "null"
            ],
            "format": "binary"
          }
        }
      },
...
      "ProfileUpdateRequest": {
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255
          }
        }
      },
...
      "ProfileDeleteRequest": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
...
      "AccountRole": {
        "type": "string",
        "enum": [
          "general_admin",
          "league_admin",
          "member",
          "guest"
        ]
      },
...
      "Branding": {
        "type": "object",
        "required": [
          "logo_url",
          "favicon_url",
          "favicon_type",
          "has_custom_logo",
          "has_custom_favicon",
          "updated_at"
        ],
        "properties": {
          "logo_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "favicon_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "favicon_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "has_custom_logo": {
            "type": "boolean"
          },
          "has_custom_favicon": {
            "type": "boolean"
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "CommandCenterSettingsPayload": {
        "type": "object",
        "required": [
          "branding"
        ],
        "properties": {
          "branding": {
            "$ref": "#/components/schemas/Branding"
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "name",
          "document_id",
          "phone",
          "address",
          "email",
          "avatar",
          "account_role",
          "account_role_label",
          "is_general_admin",
          "active_league_id",
          "email_verified_at",
          "invited_at",
          "onboarded_at",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "document_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "address": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "avatar": {
            "type": [
              "string",
              "null"
            ]
          },
          "account_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "account_role_label": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_general_admin": {
            "type": "boolean"
          },
          "active_league_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "email_verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "invited_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "onboarded_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "ErrorBag": {
        "type": "object",
        "additionalProperties": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data",
          "errors",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/User"
          },
          "errors": {
            "type": "array",
            "items": {}
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "CommandCenterSettingsResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data",
          "errors",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CommandCenterSettingsPayload"
          },
          "errors": {
            "type": "array",
            "items": {}
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "EmptySuccessResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data",
          "errors",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "null"
          },
          "errors": {
            "type": "array",
            "items": {}
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data",
          "errors",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "null"
          },
          "errors": {
            "$ref": "#/components/schemas/ErrorBag"
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  }
}